Split the re-enumeration helpers into serial-reacquire.ts
